+++ dataprocessing.m
@@ -1,10 +1,9 @@
-clc;clear;
-name_title='SH600000-1';
+function out=dataprocessing(name_title,unit)
 tail='.csv';
 filename=[name_title,tail];
 fid=fopen(filename);
 row = 0;
-unit = 2;%control the step of time
+%unit = 1;%control the step of time
 while ~feof(fid)
     row = row + 1;%get the row of the table
     fgetl(fid);
@@ -19,7 +18,9 @@ for i=2:(row)
 end
 fclose(fid);
 time=string((time));
-a=csvread(filename,1,1,[1,1,(row-1),16]);%read the numerical data from the csv file
+a=csvread(filename,1,1);%read the numerical data from the csv file
+size_a=size(a);
+col=size_a(2);
 j=0;%to change data
 i=1;%count the number of the loop
 while 1
@@ -31,6 +32,8 @@ while 1
         temp_time_hour=time_front_hour;%latest value of minute
     else 
         if (time_front_minute+60*time_front_hour)<(temp_time_minute+unit+60*temp_time_hour)
+            %delete the previous data if in the same unit of time
+            a(j+1,2)=a(j+1,2)+a(j,2);
             time(j,:)=[];
             a(j,:)=[];
             j=j-1;
@@ -64,25 +67,49 @@ while 1
     j=j+1;i=i+1;
 end
 row=sz(1);
-for i=1:sz(1)%sum the volume
-    if i==1
-    else
-        a(i,2)=a(i-1,2)+a(i,2);
-    end
+
+%for i=1:sz(1)%sum the volume
+%    if i==1
+%    else
+%        a(i,2)=a(i-1,2)+a(i,2);
+%    end
+%end
+
+if col==16    
+    %choose only one
+    a(:,5)=[];%delete SP2
+    a(:,5)=[];%delete SP3
+    a(:,6)=[];%delete SV2
+    a(:,6)=[];%delete SV3
+    a(:,7)=[];%delete BP2
+    a(:,7)=[];%delete BP3
+    a(:,8)=[];%delete BV2
+    a(:,8)=[];%delete BV3
+elseif col==24
+    a(:,5)=[];%delete SP2
+    a(:,5)=[];%delete SP3
+    a(:,5)=[];%delete SP4
+    a(:,5)=[];%delete SP5
+    a(:,6)=[];%delete SV2
+    a(:,6)=[];%delete SV3
+    a(:,6)=[];%delete SV4
+    a(:,6)=[];%delete SV5
+    a(:,7)=[];%delete BP2
+    a(:,7)=[];%delete BP3
+    a(:,7)=[];%delete BP4
+    a(:,7)=[];%delete BP5
+    a(:,8)=[];%delete BV2
+    a(:,8)=[];%delete BV3
+    a(:,8)=[];%delete BV4
+    a(:,8)=[];%delete BV5
 end
-%choose only one
-a(:,5)=[];%delete SP2
-a(:,5)=[];%delete SP3
-a(:,6)=[];%delete SV2
-a(:,6)=[];%delete SV3
-a(:,7)=[];%delete BP2
-a(:,7)=[];%delete BP3
-a(:,8)=[];%delete BV2
-a(:,8)=[];%delete BV3
 for i=1:row
     time_cut=strsplit(time(i,1),':');
     time(i)=time_cut{1,1}+":"+time_cut{1,2};
 end
-title={'Time','Price','Volume','Amount','SP1','SV1','BP1','BV1','isBuy'};
-result_table=table(time,a(:,1),a(:,2),a(:,3),a(:,4),a(:,5),a(:,6),a(:,7),a(:,8),'VariableNames',title);
+%delete amount
+a(:,3)=[];
+title={'Time','Price','Volume','SP1','SV1','BP1','BV1','isBuy'};
+result_table=table(time,a(:,1),a(:,2),a(:,3),a(:,4),a(:,5),a(:,6),a(:,7),'VariableNames',title);
 writetable(result_table,[name_title,'-processed',tail]);
+end